requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
python-dotenv>=1.0
# Google APIs
google-auth>=2.28
google-auth-oauthlib>=1.1
google-api-python-client>=2.120
# GitHub
PyGithub>=2.3
# Slack
slack_sdk>=3.27
# Vector & LLM tools
pinecone-client>=3.0
langchain>=0.1.0
openai>=1.11
cohere>=4.27
# Misc
tqdm>=4.66
